requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0,<2.0.0  # PyTorch requires NumPy < 2.0
pillow>=9.5.0

# Data processing
scikit-learn>=1.3.0

# Configuration
pyyaml>=6.0

# Logging and visualization
tensorboard>=2.13.0

# Model architectures (optional, for additional models)
timm>=0.9.0

# Development dependencies (optional)
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.3.0
# pytest>=7.3.0